gk: refactor, remove `AyaProducer` and related classes
