+++ cli/build.gradle.kts
@@ -9,7 +9,6 @@ CommonTasks.nativeImageConfig(project)
 
 dependencies {
   api(project(":base"))
-  api(project(":parser"))
   api(project(":parser-migrate"))
   api(project(":tools-repl"))
   val deps: java.util.Properties by rootProject.ext
