--- cli/src/main/java/org/aya/cli/parse/AyaProducer.java
@@ -1,877 +1,0 @@
-// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
-// Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.parse;
-
-import kala.collection.Seq;
-import kala.collection.SeqLike;
-import kala.collection.SeqView;
-import kala.collection.immutable.ImmutableSeq;
-import kala.collection.mutable.MutableSinglyLinkedList;
-import kala.control.Either;
-import kala.control.Option;
-import kala.function.BooleanFunction;
-import kala.tuple.Tuple;
-import kala.tuple.Tuple2;
-import kala.value.MutableValue;
-import org.antlr.v4.runtime.ParserRuleContext;
-import org.antlr.v4.runtime.tree.ParseTree;
-import org.antlr.v4.runtime.tree.TerminalNode;
-import org.aya.concrete.Expr;
-import org.aya.concrete.Pattern;
-import org.aya.concrete.error.BadCounterexampleWarn;
-import org.aya.concrete.error.BadModifierWarn;
-import org.aya.concrete.error.ParseError;
-import org.aya.concrete.remark.Remark;
-import org.aya.concrete.stmt.*;
-import org.aya.core.term.FormTerm;
-import org.aya.generic.Constants;
-import org.aya.generic.Modifier;
-import org.aya.generic.util.InternalException;
-import org.aya.parser.AyaParser;
-import org.aya.pretty.doc.Doc;
-import org.aya.ref.LocalVar;
-import org.aya.util.StringEscapeUtil;
-import org.aya.util.binop.Assoc;
-import org.aya.util.binop.OpDecl;
-import org.aya.util.error.SourceFile;
-import org.aya.util.error.SourcePos;
-import org.aya.util.error.WithPos;
-import org.aya.util.reporter.Reporter;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import java.util.EnumSet;
-import java.util.List;
-import java.util.function.BiFunction;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-/**
- * The ANTLR4 concrete syntax producer.
- * RIP: 2020-10-28 ~ 2022-09-30
- * I will miss you!
- * <p>
- * TODO: remove after REPL parser migration.
- *
- * @author ice1000, kiva
- */
-public record AyaProducer(
-  @NotNull Either<SourceFile, SourcePos> source,
-  @NotNull Reporter reporter
-) {
-  public Either<ImmutableSeq<Stmt>, Expr> visitRepl(AyaParser.ReplContext ctx) {
-    var expr = ctx.expr();
-    if (expr != null) return Either.right(visitExpr(expr));
-    return Either.left(Seq.wrapJava(ctx.stmt()).flatMap(this::visitStmt));
-  }
-
-  public ImmutableSeq<Stmt> visitProgram(AyaParser.ProgramContext ctx) {
-    return Seq.wrapJava(ctx.stmt()).flatMap(this::visitStmt);
-  }
-
-  public TeleDecl.PrimDecl visitPrimDecl(AyaParser.PrimDeclContext ctx) {
-    var id = ctx.weakId();
-    var name = id.getText();
-    var sourcePos = sourcePosOf(id);
-    var type = ctx.type();
-    return new TeleDecl.PrimDecl(
-      sourcePos,
-      sourcePosOf(ctx),
-      name,
-      visitTelescope(ctx.tele()),
-      type == null ? new Expr.ErrorExpr(sourcePos, Doc.plain("missing result")) : visitType(type)
-    );
-  }
-
-  public @NotNull SeqLike<Stmt> visitStmt(AyaParser.StmtContext ctx) {
-    var importCmd = ctx.importCmd();
-    if (importCmd != null) return ImmutableSeq.of(visitImportCmd(importCmd));
-    var mod = ctx.module();
-    if (mod != null) return ImmutableSeq.of(visitModule(mod));
-    var openCmd = ctx.openCmd();
-    if (openCmd != null) return visitOpenCmd(openCmd);
-    var decl = ctx.decl();
-    if (decl != null) {
-      var result = visitDecl(decl);
-      var stmts = result._2.view().prepended(result._1);
-      if (result._1 instanceof Decl.TopLevel top && top.personality() == Decl.Personality.COUNTEREXAMPLE) {
-        var stmtOption = result._2.firstOption(stmt -> !(stmt instanceof Decl));
-        if (stmtOption.isDefined()) reporter.report(new BadCounterexampleWarn(stmtOption.get()));
-        return stmts.<Stmt>filterIsInstance(Decl.class).toImmutableSeq();
-      }
-      return stmts;
-    }
-    var generalize = ctx.generalize();
-    if (generalize != null) return ImmutableSeq.of(visitGeneralize(generalize));
-    var remark = ctx.remark();
-    if (remark != null) return ImmutableSeq.of(visitRemark(remark));
-    return unreachable(ctx);
-  }
-
-  @NotNull private Remark visitRemark(AyaParser.RemarkContext remark) {
-    var sb = new StringBuilder();
-    for (var docComment : remark.DOC_COMMENT()) {
-      sb.append(docComment.getText().substring(3)).append("\n");
-    }
-    return Remark.make(sb.toString(), sourcePosOf(remark), new AyaParserImpl(reporter));
-  }
-
-  public Generalize visitGeneralize(AyaParser.GeneralizeContext ctx) {
-    return new Generalize(sourcePosOf(ctx), visitIds(ctx.ids())
-      .map(id -> new GeneralizedVar(id.data(), id.sourcePos()))
-      .collect(ImmutableSeq.factory()), visitType(ctx.type()));
-  }
-
-  public @NotNull BindBlock visitBind(AyaParser.BindBlockContext ctx) {
-    return new BindBlock(sourcePosOf(ctx), MutableValue.create(),
-      ctx.loosers().stream().flatMap(c -> c.qualifiedId().stream().map(this::visitQualifiedId)).collect(ImmutableSeq.factory()),
-      ctx.tighters().stream().flatMap(c1 -> c1.qualifiedId().stream().map(this::visitQualifiedId)).collect(ImmutableSeq.factory()),
-      MutableValue.create(), MutableValue.create());
-  }
-
-  private <T> T unreachable(ParserRuleContext ctx) {
-    throw new InternalException(ctx.getClass() + ": " + ctx.getText());
-  }
-
-  public @NotNull Tuple2<? extends Decl, ImmutableSeq<Stmt>> visitDecl(AyaParser.DeclContext ctx) {
-    var accessibility = ctx.PRIVATE() == null ? Stmt.Accessibility.Public : Stmt.Accessibility.Private;
-    var fnDecl = ctx.fnDecl();
-    if (fnDecl != null) return Tuple.of(visitFnDecl(fnDecl, accessibility), ImmutableSeq.empty());
-    var dataDecl = ctx.dataDecl();
-    if (dataDecl != null) return visitDataDecl(dataDecl, accessibility);
-    var structDecl = ctx.structDecl();
-    if (structDecl != null) return visitStructDecl(structDecl, accessibility);
-    var primDecl = ctx.primDecl();
-    if (primDecl != null) return Tuple.of(visitPrimDecl(primDecl), ImmutableSeq.empty());
-    return unreachable(ctx);
-  }
-
-  public Tuple2<@NotNull WithPos<String>, OpDecl.@Nullable OpInfo> visitDeclNameOrInfix(@NotNull AyaParser.DeclNameOrInfixContext ctx) {
-    var assoc = ctx.assoc();
-    var id = ctx.weakId();
-    var txt = id.getText();
-    var pos = sourcePosOf(id);
-    if (assoc == null) return Tuple.of(new WithPos<>(pos, txt), null);
-    var infix = new OpDecl.OpInfo(txt, visitAssoc(assoc));
-    return Tuple.of(new WithPos<>(pos, infix.name()), infix);
-  }
-
-  private @NotNull Assoc visitAssoc(@NotNull AyaParser.AssocContext assoc) {
-    if (assoc.INFIX() != null) return Assoc.Infix;
-    if (assoc.INFIXL() != null) return Assoc.InfixL;
-    if (assoc.INFIXR() != null) return Assoc.InfixR;
-    if (assoc.FIXL() != null) return Assoc.FixL;
-    if (assoc.FIXR() != null) return Assoc.FixR;
-    throw new InternalException("Unknown assoc: " + assoc.getText());
-  }
-
-  public TeleDecl.@NotNull FnDecl visitFnDecl(AyaParser.FnDeclContext ctx, Stmt.Accessibility accessibility) {
-    var personality = visitSampleModifiers(ctx.sampleModifiers());
-    var modifiers = Seq.from(ctx.fnModifiers()).view()
-      .map(fn -> Tuple.of(fn, visitFnModifiers(fn)))
-      .toImmutableSeq();
-    var inline = modifiers.find(t -> t._2 == Modifier.Inline);
-    var opaque = modifiers.find(t -> t._2 == Modifier.Opaque);
-    if (inline.isDefined() && opaque.isDefined()) {
-      var gunpowder = inline.get();
-      reporter.report(new BadModifierWarn(sourcePosOf(gunpowder._1), gunpowder._2));
-    }
-
-    var tele = visitTelescope(ctx.tele());
-    var bind = ctx.bindBlock();
-    var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
-
-    var dynamite = visitFnBody(ctx.fnBody());
-    if (dynamite.isRight() && inline.isDefined()) {
-      var gelatin = inline.get();
-      reporter.report(new BadModifierWarn(sourcePosOf(gelatin._1), gelatin._2));
-    }
-    return new TeleDecl.FnDecl(
-      nameOrInfix._1.sourcePos(),
-      sourcePosOf(ctx),
-      personality == Decl.Personality.NORMAL ? accessibility : Stmt.Accessibility.Private,
-      modifiers.map(Tuple2::getValue).collect(Collectors.toCollection(
-        () -> EnumSet.noneOf(Modifier.class))),
-      nameOrInfix._2,
-      nameOrInfix._1.data(),
-      tele,
-      type(ctx.type(), sourcePosOf(ctx)),
-      dynamite,
-      bind == null ? BindBlock.EMPTY : visitBind(bind),
-      personality
-    );
-  }
-
-  public @NotNull Decl.Personality visitSampleModifiers(AyaParser.SampleModifiersContext ctx) {
-    if (ctx == null) return Decl.Personality.NORMAL;
-    if (ctx.EXAMPLE() != null) return Decl.Personality.EXAMPLE;
-    return Decl.Personality.COUNTEREXAMPLE;
-  }
-
-  public @NotNull ImmutableSeq<Expr.@NotNull Param> visitTelescope(List<AyaParser.TeleContext> telescope) {
-    return Seq.wrapJava(telescope).flatMap(t -> visitTele(t, false));
-  }
-
-  public @NotNull ImmutableSeq<Expr.@NotNull Param> visitLamTelescope(List<AyaParser.TeleContext> telescope) {
-    return Seq.wrapJava(telescope).flatMap(t -> visitTele(t, true));
-  }
-
-  public @NotNull ImmutableSeq<Expr.@NotNull Param> visitForallTelescope(List<AyaParser.TeleContext> telescope) {
-    return Seq.wrapJava(telescope).flatMap(t -> visitTele(t, true));
-  }
-
-  public @NotNull Either<Expr, ImmutableSeq<Pattern.Clause>> visitFnBody(AyaParser.FnBodyContext ctx) {
-    var expr = ctx.expr();
-    if (expr != null) return Either.left(visitExpr(expr));
-    return Either.right(Seq.wrapJava(ctx.clause()).map(this::visitClause));
-  }
-
-  public QualifiedID visitQualifiedId(AyaParser.QualifiedIdContext ctx) {
-    return new QualifiedID(sourcePosOf(ctx),
-      ctx.weakId().stream().map(ParseTree::getText)
-        .collect(ImmutableSeq.factory()));
-  }
-
-  public @NotNull Expr visitLiteral(AyaParser.LiteralContext ctx) {
-    var pos = sourcePosOf(ctx);
-    if (ctx.CALM_FACE() != null) return new Expr.HoleExpr(pos, false, null);
-    var id = ctx.qualifiedId();
-    if (id != null) return new Expr.UnresolvedExpr(pos, visitQualifiedId(id));
-    if (ctx.TYPE() != null) return new Expr.RawSortExpr(pos, FormTerm.SortKind.Type);
-    if (ctx.SET_KW() != null) return new Expr.RawSortExpr(pos, FormTerm.SortKind.Set);
-    if (ctx.PROP() != null) return new Expr.RawSortExpr(pos, FormTerm.SortKind.Prop);
-    if (ctx.ISET() != null) return new Expr.RawSortExpr(pos, FormTerm.SortKind.ISet);
-    if (ctx.LGOAL() != null) {
-      var fillingExpr = ctx.expr();
-      var filling = fillingExpr == null ? null : visitExpr(fillingExpr);
-      return new Expr.HoleExpr(pos, true, filling);
-    }
-    var number = ctx.NUMBER();
-    if (number != null) try {
-      return new Expr.LitIntExpr(pos, Integer.parseInt(number.getText()));
-    } catch (NumberFormatException ignored) {
-      reporter.report(new ParseError(sourcePosOf(number), "Unsupported integer literal `" + number.getText() + "`"));
-      throw new ParsingInterruptedException();
-    }
-    var string = ctx.STRING();
-    if (string != null) {
-      var content = string.getText().substring(1, string.getText().length() - 1);
-      return new Expr.LitStringExpr(pos, StringEscapeUtil.escapeStringCharacters(content));
-    }
-    return unreachable(ctx);
-  }
-
-  private @NotNull LocalVar visitParamLiteral(AyaParser.LiteralContext ctx) {
-    var idCtx = ctx.qualifiedId();
-    var id = visitQualifiedId(idCtx);
-    return new LocalVar(id.justName(), sourcePosOf(idCtx));
-  }
-
-  public @NotNull ImmutableSeq<Expr.@NotNull Param> visitTele(AyaParser.TeleContext ctx, boolean isParamLiteral) {
-    var literal = ctx.literal();
-    if (literal != null) {
-      var pos = sourcePosOf(ctx);
-      return ImmutableSeq.of(isParamLiteral
-        ? new Expr.Param(pos, visitParamLiteral(literal), type(null, pos), true)
-        : new Expr.Param(pos, Constants.randomlyNamed(pos), visitLiteral(literal), true)
-      );
-    }
-    var teleBinder = ctx.teleBinder();
-    var teleMaybeTypedExpr = ctx.teleMaybeTypedExpr();
-    if (teleBinder != null) {
-      var type = teleBinder.expr();
-      if (type != null) {
-        var pos = sourcePosOf(ctx);
-        return ImmutableSeq.of(new Expr.Param(pos, Constants.randomlyNamed(pos), visitExpr(type), true));
-      }
-      teleMaybeTypedExpr = teleBinder.teleMaybeTypedExpr();
-    }
-    if (ctx.LPAREN() != null) return visitTeleMaybeTypedExpr(teleMaybeTypedExpr).apply(true);
-    if (ctx.LBRACE() != null) return visitTeleMaybeTypedExpr(teleMaybeTypedExpr).apply(false);
-    return unreachable(ctx);
-  }
-
-  public @NotNull BooleanFunction<ImmutableSeq<Expr.Param>>
-  visitTeleMaybeTypedExpr(AyaParser.TeleMaybeTypedExprContext ctx) {
-    var ids = ctx.ids();
-    var type = type(ctx.type(), sourcePosOf(ids));
-    return explicit -> visitIds(ids)
-      .map(v -> new Expr.Param(v.sourcePos(), LocalVar.from(v), type, explicit))
-      .collect(ImmutableSeq.factory());
-  }
-
-  public @NotNull Expr visitExpr(AyaParser.ExprContext ctx) {
-    return switch (ctx) {
-      case AyaParser.SingleContext sin -> visitAtom(sin.atom());
-      case AyaParser.AppContext app -> {
-        var head = new Expr.NamedArg(true, visitExpr(app.expr()));
-        var tail = app.argument().stream()
-          .map(this::visitArgument)
-          .collect(MutableSinglyLinkedList.factory());
-        tail.push(head);
-        yield new Expr.BinOpSeq(sourcePosOf(app), tail.toImmutableSeq());
-      }
-      case AyaParser.ProjContext proj -> buildProj(sourcePosOf(proj), visitExpr(proj.expr()), proj.projFix());
-      case AyaParser.PiContext pi -> buildPi(
-        sourcePosOf(pi), false,
-        visitTelescope(pi.tele()).view(),
-        visitExpr(pi.expr()));
-      case AyaParser.SigmaContext sig -> new Expr.SigmaExpr(
-        sourcePosOf(sig), false,
-        visitTelescope(sig.tele()).appended(new Expr.Param(visitExpr(sig.expr()).sourcePos(), LocalVar.IGNORED, visitExpr(sig.expr()), true)));
-      case AyaParser.LamContext lam -> {
-        Expr result;
-        var bodyExpr = lam.expr();
-        if (bodyExpr == null) {
-          var impliesToken = lam.IMPLIES();
-          var bodyHolePos = impliesToken == null ? sourcePosOf(lam) : sourcePosOf(impliesToken);
-          result = new Expr.HoleExpr(bodyHolePos, false, null);
-        } else result = visitExpr(bodyExpr);
-        yield buildLam(sourcePosOf(lam), visitLamTelescope(lam.tele()).view(), result);
-      }
-      case AyaParser.ArrContext arr -> {
-        var expr0 = arr.expr(0);
-        var to = visitExpr(arr.expr(1));
-        var pos = sourcePosOf(expr0);
-        var param = new Expr.Param(pos, Constants.randomlyNamed(pos), visitExpr(expr0), true);
-        yield new Expr.PiExpr(sourcePosOf(arr), false, param, to);
-      }
-      case AyaParser.NewContext n -> new Expr.NewExpr(
-        sourcePosOf(n), visitExpr(n.expr()),
-        Option.ofNullable(n.newBody()).map(b -> Seq.wrapJava(b.newArg()).map(this::visitField))
-          .getOrDefault(ImmutableSeq.empty()));
-      case AyaParser.ForallContext forall -> buildPi(
-        sourcePosOf(forall), false,
-        visitForallTelescope(forall.tele()).view(),
-        visitExpr(forall.expr()));
-      case AyaParser.DoContext doCtx -> visitDo(doCtx);
-      case AyaParser.IdiomContext idmCtx -> {
-        if (idmCtx.idiomBlock() == null)
-          yield Constants.alternativeEmpty(sourcePosOf(idmCtx));
-        yield visitIdiomBlock(idmCtx.idiomBlock());
-      }
-      case AyaParser.ArrayContext arrCtx -> {
-        if (arrCtx.arrayBlock() == null)
-          yield Constants.listNil(sourcePosOf(arrCtx));
-        else if (arrCtx.arrayBlock().BAR() == null)
-          yield visitArray(arrCtx.arrayBlock());
-        else
-          yield visitListComprehension(arrCtx.arrayBlock());
-      }
-      case AyaParser.PartElContext elCtx -> visitPartial(elCtx.partial(), elCtx);
-      case AyaParser.PathContext path -> new Expr.Path(
-        sourcePosOf(path),
-        Seq.wrapJava(path.weakId()).map(w -> new LocalVar(w.getText(), sourcePosOf(w))),
-        visitExpr(path.expr()),
-        visitPartial(path.partial(), path)
-      );
-      // TODO: match
-      default -> throw new UnsupportedOperationException("TODO: " + ctx.getClass());
-    };
-  }
-
-  private @NotNull Expr.PartEl visitPartial(
-    @Nullable AyaParser.PartialContext ctx,
-    @NotNull ParserRuleContext fallback
-  ) {
-    if (ctx == null) return new Expr.PartEl(sourcePosOf(fallback), ImmutableSeq.empty());
-    return new Expr.PartEl(sourcePosOf(ctx), Seq.wrapJava(ctx.subSystem()).map(sys -> {
-      var exprs = Seq.wrapJava(sys.expr()).map(this::visitExpr);
-      return Tuple.of(exprs.get(0), exprs.get(1));
-    }));
-  }
-
-
-  private @NotNull Expr visitListComprehension(AyaParser.ArrayBlockContext ctx) {
-    var firstExpr = new Expr.NamedArg(true, visitExpr(ctx.expr()));
-    var pure = new Expr.NamedArg(true, Constants.functorPure(firstExpr.sourcePos()));
-    var returnedExpr = new Expr.BinOpSeq(firstExpr.sourcePos(), ImmutableSeq.of(pure, firstExpr));
-
-    return Seq.wrapJava(ctx.listComp().doBindingExpr())
-      .foldRight(returnedExpr, (doBindingCtx, accExpr) -> {
-        var bindOp = new Expr.NamedArg(true, Constants.monadBind(sourcePosOf(doBindingCtx.LARROW())));
-        var pos = sourcePosOf(doBindingCtx);
-        var param = new Expr.Param(sourcePosOf(doBindingCtx.weakId()),
-          new LocalVar(doBindingCtx.weakId().getText()), true);
-
-        var rhs = new Expr.NamedArg(true, new Expr.LamExpr(pos, param, accExpr));
-        var lhs = new Expr.NamedArg(true, visitExpr(doBindingCtx.expr()));
-        var seq = ImmutableSeq.of(lhs, bindOp, rhs);
-        return new Expr.BinOpSeq(pos, seq);
-      });
-  }
-
-  private @NotNull Expr visitArray(AyaParser.ArrayBlockContext ctx) {
-    var consArg = new Expr.NamedArg(true, Constants.listCons(sourcePosOf(ctx)));
-    var nilArg = new Expr.NamedArg(true, Constants.listNil(sourcePosOf(ctx)));
-    var args = Seq.wrapJava(ctx.exprList().expr()).view()
-      .map(exprCtx -> new Expr.NamedArg(true, visitExpr(exprCtx)))
-      .flatMap(expr -> ImmutableSeq.of(expr, consArg))
-      .appended(nilArg)
-      .toImmutableSeq();
-    return new Expr.BinOpSeq(sourcePosOf(ctx), args);
-  }
-
-  /**
-   * Warning: the parser cannot enforce left associativity at this stage
-   */
-  private @NotNull Expr visitUnBarredIdiom(@NotNull SourcePos pos, List<AyaParser.ExprContext> ctxs) {
-    var apSeq = buildApSeq(pos, ctxs);
-    var pure = Constants.functorPure(apSeq.first().sourcePos());
-    var pureFirst = new Expr.NamedArg(true, new Expr.AppExpr(pure.sourcePos(), pure, apSeq.first()));
-    return new Expr.BinOpSeq(pos, apSeq.drop(1).prepended(pureFirst).toImmutableSeq());
-  }
-
-  private @NotNull Expr visitIdiomBlock(AyaParser.IdiomBlockContext ctx) {
-    var lastIdiom = visitUnBarredIdiom(sourcePosOf(ctx), ctx.expr());
-
-    var barredExpr = Seq.wrapJava(ctx.barredExpr());
-    if (barredExpr.isEmpty())
-      return lastIdiom;
-
-    var appSeq = barredExpr.view()
-      .flatMap(barredExprCtx -> {
-        var unBarred = visitUnBarredIdiom(sourcePosOf(barredExprCtx), barredExprCtx.expr());
-        var unBarredArg = new Expr.NamedArg(true, unBarred);
-        var orArg = new Expr.NamedArg(true, Constants.alternativeOr(sourcePosOf(barredExprCtx.BAR())));
-        return ImmutableSeq.of(unBarredArg, orArg);
-      })
-      .appended(new Expr.NamedArg(true, lastIdiom))
-      .toImmutableSeq();
-    return new Expr.BinOpSeq(sourcePosOf(ctx), appSeq);
-  }
-
-  private @NotNull SeqView<Expr.NamedArg> buildApSeq(@NotNull SourcePos pos, @NotNull List<AyaParser.ExprContext> exprs) {
-    var ap = new Expr.NamedArg(true, Constants.applicativeApp(pos));
-    return Seq.wrapJava(exprs).view()
-      .map(expr -> new Expr.NamedArg(true, visitExpr(expr)))
-      .flatMap(arg -> ImmutableSeq.of(ap, arg))
-      .drop(1);
-  }
-
-  private @NotNull Expr visitDo(AyaParser.DoContext ctx) {
-    var doBlockExprCtxs = Seq.wrapJava(ctx.doBlock().doBlockExpr());
-    var lastExprCtx = doBlockExprCtxs.last();
-    if (lastExprCtx.doBindingExpr() != null) {
-      reporter.report(new ParseError(sourcePosOf(lastExprCtx),
-        "last expression in a do block cannot be a bind expression"));
-      throw new ParsingInterruptedException();
-    }
-
-    var lastExpr = visitExpr(lastExprCtx.expr());
-    return doBlockExprCtxs.view().dropLast(1).foldRight(lastExpr, (doCtx, accExpr) -> {
-      var doBinding = doCtx.doBindingExpr();
-
-      var lArrow = doBinding != null ? doBinding.LARROW() : null;
-      var pos = lArrow != null ? sourcePosOf(lArrow) : sourcePosOf(doCtx);
-      var bindOp = new Expr.NamedArg(true, Constants.monadBind(pos));
-
-      var sourcePos = sourcePosOf(doCtx);
-      var param = doBinding != null
-        ? new Expr.Param(sourcePosOf(doBinding.weakId()), new LocalVar(doBinding.weakId().getText()), true)
-        : new Expr.Param(sourcePosOf(doCtx), LocalVar.IGNORED, true);
-
-      var rhs = new Expr.NamedArg(true, new Expr.LamExpr(sourcePos, param, accExpr));
-      var lhs = doBinding != null
-        ? new Expr.NamedArg(true, visitExpr(doBinding.expr()))
-        : new Expr.NamedArg(true, visitExpr(doCtx.expr()));
-
-      var seq = ImmutableSeq.of(lhs, bindOp, rhs);
-      return new Expr.BinOpSeq(sourcePos, seq);
-    });
-  }
-
-  private @NotNull Expr.Field visitField(AyaParser.NewArgContext na) {
-    var weakId = na.weakId();
-    return new Expr.Field(new WithPos<>(sourcePosOf(weakId), weakId.getText()), visitIds(na.ids())
-      .map(t -> new WithPos<>(t.sourcePos(), LocalVar.from(t)))
-      .collect(ImmutableSeq.factory()), visitExpr(na.expr()), MutableValue.create());
-  }
-
-  public @NotNull Expr visitAtom(AyaParser.AtomContext ctx) {
-    var literal = ctx.literal();
-    if (literal != null) {
-      var expr = visitLiteral(literal);
-      var lifts = ctx.ULIFT().size();
-      return lifts > 0 ? new Expr.LiftExpr(sourcePosOf(ctx), expr, lifts) : expr;
-    }
-
-    var expr = ctx.exprList().expr();
-    if (expr.size() == 1) return newBinOPScope(visitExpr(expr.get(0)));
-    // if (expr.size() == 1) return visitExpr(expr.get(0));
-    return new Expr.TupExpr(
-      sourcePosOf(ctx),
-      expr.stream()
-        .map(this::visitExpr)
-        .collect(ImmutableSeq.factory())
-    );
-  }
-
-  public @NotNull Expr.NamedArg visitArgument(AyaParser.ArgumentContext ctx) {
-    var atom = ctx.atom();
-    if (atom != null) {
-      var fixes = ctx.projFix();
-      var expr = visitAtom(atom);
-      var projected = Seq.wrapJava(fixes)
-        .foldLeft(Tuple.of(sourcePosOf(ctx), expr),
-          (acc, proj) -> Tuple.of(acc._2.sourcePos(), buildProj(acc._1, acc._2, proj)))
-        ._2;
-      return new Expr.NamedArg(true, projected);
-    }
-    // assert ctx.LBRACE() != null;
-    var id = ctx.weakId();
-    if (id != null) return new Expr.NamedArg(false, id.getText(), visitExpr(ctx.expr()));
-    var items = Seq.wrapJava(ctx.exprList().expr()).map(this::visitExpr);
-    if (items.sizeEquals(1)) return new Expr.NamedArg(false, newBinOPScope(items.first()));
-    var tupExpr = new Expr.TupExpr(sourcePosOf(ctx), items);
-    return new Expr.NamedArg(false, tupExpr);
-  }
-
-  /**
-   * [kiva]: make `(expr)` into a new BinOP parser scope
-   * so the `f (+)` becomes passing `+` as an argument to function `f`.
-   * this should be a workaround?
-   * see base/src/test/resources/success/binop.aya
-   */
-  public @NotNull Expr newBinOPScope(@NotNull Expr expr) {
-    return new Expr.BinOpSeq(expr.sourcePos(),
-      ImmutableSeq.of(new Expr.NamedArg(true, expr)));
-  }
-
-  public @NotNull Pattern newBinOPScope(@NotNull Pattern expr, @Nullable LocalVar as) {
-    return new Pattern.BinOpSeq(expr.sourcePos(),
-      ImmutableSeq.of(expr), as, expr.explicit());
-  }
-
-  public static @NotNull Expr buildLam(SourcePos sourcePos, SeqView<Expr.Param> params, Expr body) {
-    if (params.isEmpty()) return body;
-    var drop = params.drop(1);
-    return new Expr.LamExpr(
-      sourcePos, params.first(),
-      buildLam(AntlrUtil.sourcePosForSubExpr(sourcePos.file(),
-        drop.map(Expr.Param::sourcePos), body.sourcePos()), drop, body));
-  }
-
-  public static @NotNull Expr buildPi(
-    SourcePos sourcePos, boolean co,
-    SeqView<Expr.Param> params, Expr body
-  ) {
-    if (params.isEmpty()) return body;
-    var drop = params.drop(1);
-    return new Expr.PiExpr(
-      sourcePos, co, params.first(),
-      buildPi(AntlrUtil.sourcePosForSubExpr(sourcePos.file(),
-        drop.map(Expr.Param::sourcePos), body.sourcePos()), co, drop, body));
-  }
-
-  private Expr.@NotNull ProjExpr buildProj(
-    @NotNull SourcePos sourcePos, @NotNull Expr projectee,
-    @NotNull AyaParser.ProjFixContext fix
-  ) {
-    var number = fix.NUMBER();
-    return new Expr.ProjExpr(
-      sourcePos,
-      projectee,
-      number != null
-        ? Either.left(Integer.parseInt(number.getText()))
-        : Either.right(visitQualifiedId(fix.qualifiedId()))
-    );
-  }
-
-  public @NotNull Tuple2<TeleDecl.DataDecl, ImmutableSeq<Stmt>>
-  visitDataDecl(AyaParser.DataDeclContext ctx, Stmt.Accessibility accessibility) {
-    var personality = visitSampleModifiers(ctx.sampleModifiers());
-    var bind = ctx.bindBlock();
-    var openAccessibility = ctx.PUBLIC() != null ? Stmt.Accessibility.Public : Stmt.Accessibility.Private;
-    var body = ctx.dataBody().stream().map(this::visitDataBody).collect(ImmutableSeq.factory());
-    var tele = visitTelescope(ctx.tele());
-    var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
-    var data = new TeleDecl.DataDecl(
-      nameOrInfix._1.sourcePos(),
-      sourcePosOf(ctx),
-      personality == Decl.Personality.NORMAL ? accessibility : Stmt.Accessibility.Private,
-      nameOrInfix._2,
-      nameOrInfix._1.data(),
-      tele,
-      type(ctx.type(), sourcePosOf(ctx)),
-      body,
-      bind == null ? BindBlock.EMPTY : visitBind(bind),
-      personality
-    );
-    return Tuple.of(data, ctx.OPEN() == null ? ImmutableSeq.empty() : ImmutableSeq.of(
-      new Command.Open(
-        sourcePosOf(ctx.OPEN()),
-        openAccessibility,
-        new QualifiedID(sourcePosOf(ctx), nameOrInfix._1.data()),
-        UseHide.EMPTY,
-        personality == Decl.Personality.EXAMPLE
-      )
-    ));
-  }
-
-  public @NotNull Expr type(@Nullable AyaParser.TypeContext typeCtx, SourcePos sourcePos) {
-    return typeCtx == null
-      ? new Expr.HoleExpr(sourcePos, false, null)
-      : visitType(typeCtx);
-  }
-
-  private @NotNull TeleDecl.DataCtor visitDataBody(AyaParser.DataBodyContext ctx) {
-    if (ctx instanceof AyaParser.DataCtorsContext dcc) return visitDataCtor(ImmutableSeq.empty(), dcc.dataCtor());
-    if (ctx instanceof AyaParser.DataClausesContext dcc) return visitDataCtorClause(dcc.dataCtorClause());
-    return unreachable(ctx);
-  }
-
-  public TeleDecl.DataCtor visitDataCtor(@NotNull ImmutableSeq<Pattern> patterns, AyaParser.DataCtorContext ctx) {
-    var tele = visitTelescope(ctx.tele());
-    var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
-    var bind = ctx.bindBlock();
-    return new TeleDecl.DataCtor(
-      nameOrInfix._1.sourcePos(),
-      sourcePosOf(ctx),
-      nameOrInfix._2,
-      nameOrInfix._1.data(),
-      tele,
-      visitClauses(ctx.clauses()),
-      patterns,
-      ctx.COERCE() != null,
-      bind == null ? BindBlock.EMPTY : visitBind(bind)
-    );
-  }
-
-  public ImmutableSeq<Pattern.Clause> visitClauses(@Nullable AyaParser.ClausesContext ctx) {
-    if (ctx == null) return ImmutableSeq.empty();
-    return Seq.wrapJava(ctx.clause()).map(this::visitClause);
-  }
-
-  public @NotNull TeleDecl.DataCtor visitDataCtorClause(AyaParser.DataCtorClauseContext ctx) {
-    return visitDataCtor(visitPatterns(ctx.patterns()), ctx.dataCtor());
-  }
-
-  public @NotNull Pattern visitPattern(AyaParser.PatternContext ctx) {
-    return visitAtomPatterns(ctx.atomPatterns()).apply(true, null);
-  }
-
-  public BiFunction<Boolean, LocalVar, Pattern> visitAtomPatterns(@NotNull AyaParser.AtomPatternsContext ctx) {
-    var atoms = ctx.atomPattern().stream()
-      .map(this::visitAtomPattern)
-      .collect(ImmutableSeq.factory());
-    if (atoms.sizeEquals(1)) return (ex, as) -> newBinOPScope(atoms.first().apply(ex), as);
-
-    return (ex, as) -> new Pattern.BinOpSeq(
-      sourcePosOf(ctx),
-      atoms.view().map(p -> p.apply(true)).toImmutableSeq(),
-      as, ex
-    );
-  }
-
-  public @NotNull BooleanFunction<Pattern> visitAtomPattern(AyaParser.AtomPatternContext ctx) {
-    var sourcePos = sourcePosOf(ctx);
-    if (ctx.LPAREN() != null || ctx.LBRACE() != null) {
-      var forceEx = ctx.LPAREN() != null;
-      var patterns = ctx.patterns();
-      if (patterns == null) return ex -> new Pattern.Absurd(sourcePos, ex);
-      var id = ctx.weakId();
-      var as = id != null ? new LocalVar(id.getText(), sourcePosOf(id)) : null;
-      var tupElem = Seq.wrapJava(patterns.pattern())
-        .map(t -> visitAtomPatterns(t.atomPatterns()));
-      return tupElem.sizeEquals(1)
-        ? (ignored -> newBinOPScope(tupElem.first().apply(forceEx, as), as))
-        : (ignored -> new Pattern.Tuple(sourcePos, forceEx, tupElem.map(p -> p.apply(true, null)), as));
-    }
-    if (ctx.CALM_FACE() != null) return ex -> new Pattern.CalmFace(sourcePos, ex);
-    var number = ctx.NUMBER();
-    if (number != null) return ex -> new Pattern.Number(sourcePos, ex, Integer.parseInt(number.getText()));
-    var id = ctx.weakId();
-    if (id != null)
-      return ex -> new Pattern.Bind(sourcePos, ex, new LocalVar(id.getText(), sourcePosOf(id)), MutableValue.create());
-
-    return unreachable(ctx);
-  }
-
-  public @NotNull ImmutableSeq<Pattern> visitPatterns(AyaParser.PatternsContext ctx) {
-    return ctx.pattern().stream()
-      .map(this::visitPattern)
-      .collect(ImmutableSeq.factory());
-  }
-
-  public @NotNull Pattern.Clause visitClause(AyaParser.ClauseContext ctx) {
-    return new Pattern.Clause(sourcePosOf(ctx), visitPatterns(ctx.patterns()),
-      Option.ofNullable(ctx.expr()).map(this::visitExpr));
-  }
-
-  public @NotNull Tuple2<TeleDecl.StructDecl, ImmutableSeq<Stmt>>
-  visitStructDecl(AyaParser.StructDeclContext ctx, Stmt.Accessibility accessibility) {
-    var personality = visitSampleModifiers(ctx.sampleModifiers());
-    var bind = ctx.bindBlock();
-    var openAccessibility = ctx.PUBLIC() != null ? Stmt.Accessibility.Public : Stmt.Accessibility.Private;
-    var fields = visitFields(ctx.field());
-    var tele = visitTelescope(ctx.tele());
-    var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
-    var struct = new TeleDecl.StructDecl(
-      nameOrInfix._1.sourcePos(),
-      sourcePosOf(ctx),
-      personality == Decl.Personality.NORMAL ? accessibility : Stmt.Accessibility.Private,
-      nameOrInfix._2,
-      nameOrInfix._1.data(),
-      tele,
-      type(ctx.type(), sourcePosOf(ctx)),
-      fields,
-      bind == null ? BindBlock.EMPTY : visitBind(bind),
-      personality
-    );
-    return Tuple.of(struct, ctx.OPEN() == null ? ImmutableSeq.empty() : ImmutableSeq.of(
-      new Command.Open(
-        sourcePosOf(ctx.OPEN()),
-        openAccessibility,
-        new QualifiedID(sourcePosOf(ctx), nameOrInfix._1.data()),
-        UseHide.EMPTY,
-        personality == Decl.Personality.EXAMPLE
-      )
-    ));
-  }
-
-  private ImmutableSeq<TeleDecl.StructField> visitFields(List<AyaParser.FieldContext> field) {
-    return Seq.wrapJava(field).map(fieldCtx -> {
-      if (fieldCtx instanceof AyaParser.FieldDeclContext fieldDecl) return visitFieldDecl(fieldDecl);
-      else if (fieldCtx instanceof AyaParser.FieldImplContext fieldImpl) return visitFieldImpl(fieldImpl);
-      else throw new InternalException(fieldCtx.getClass() + " is neither FieldDecl nor FieldImpl!");
-    });
-  }
-
-  public TeleDecl.StructField visitFieldImpl(AyaParser.FieldImplContext ctx) {
-    var tele = visitTelescope(ctx.tele());
-    var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
-    var bind = ctx.bindBlock();
-    return new TeleDecl.StructField(
-      nameOrInfix._1.sourcePos(),
-      sourcePosOf(ctx),
-      nameOrInfix._2,
-      nameOrInfix._1.data(),
-      tele,
-      type(ctx.type(), sourcePosOf(ctx)),
-      Option.ofNullable(ctx.expr()).map(this::visitExpr),
-      false,
-      bind == null ? BindBlock.EMPTY : visitBind(bind)
-    );
-  }
-
-  public TeleDecl.StructField visitFieldDecl(AyaParser.FieldDeclContext ctx) {
-    var tele = visitTelescope(ctx.tele());
-    var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
-    var bind = ctx.bindBlock();
-    return new TeleDecl.StructField(
-      nameOrInfix._1.sourcePos(),
-      sourcePosOf(ctx),
-      nameOrInfix._2,
-      nameOrInfix._1.data(),
-      tele,
-      type(ctx.type(), sourcePosOf(ctx)),
-      Option.none(),
-      ctx.COERCE() != null,
-      bind == null ? BindBlock.EMPTY : visitBind(bind)
-    );
-  }
-
-  public @NotNull Expr visitType(@NotNull AyaParser.TypeContext ctx) {
-    return visitExpr(ctx.expr());
-  }
-
-  public @NotNull Stmt visitImportCmd(AyaParser.ImportCmdContext ctx) {
-    final var id = ctx.weakId();
-    return new Command.Import(
-      sourcePosOf(ctx.qualifiedId()),
-      visitQualifiedId(ctx.qualifiedId()),
-      id == null ? null : id.getText()
-    );
-  }
-
-  public @NotNull ImmutableSeq<Stmt> visitOpenCmd(AyaParser.OpenCmdContext ctx) {
-    var accessibility = ctx.PUBLIC() == null
-      ? Stmt.Accessibility.Private
-      : Stmt.Accessibility.Public;
-    var useHide = ctx.useHide();
-    var modNameCtx = ctx.qualifiedId();
-    var namePos = sourcePosOf(modNameCtx);
-    var modName = visitQualifiedId(modNameCtx);
-    var open = new Command.Open(
-      namePos,
-      accessibility,
-      modName,
-      useHide != null ? visitUseHide(useHide) : UseHide.EMPTY,
-      false
-    );
-    if (ctx.IMPORT() != null) return ImmutableSeq.of(
-      new Command.Import(namePos, modName, null),
-      open
-    );
-    else return ImmutableSeq.of(open);
-  }
-
-  public UseHide hideList(List<AyaParser.HideListContext> ctxs, UseHide.Strategy strategy) {
-    return new UseHide(
-      ctxs.stream()
-        .map(AyaParser.HideListContext::idsComma)
-        .flatMap(this::visitIdsComma)
-        .map(id -> new UseHide.Name(id, id, Assoc.Invalid, BindBlock.EMPTY))
-        .collect(ImmutableSeq.factory()),
-      strategy);
-  }
-
-  public UseHide useList(List<AyaParser.UseListContext> ctxs, UseHide.Strategy strategy) {
-    return new UseHide(ctxs.stream()
-      .map(AyaParser.UseListContext::useIdsComma)
-      .flatMap(this::visitUseIdsComma)
-      .collect(ImmutableSeq.factory()),
-      strategy);
-  }
-
-  public Stream<UseHide.Name> visitUseIdsComma(@NotNull AyaParser.UseIdsCommaContext ctx) {
-    return ctx.useId().stream().map(id -> {
-      var name = id.weakId().getText();
-      var useAs = id.useAs();
-      if (useAs == null) return new UseHide.Name(name, name, Assoc.Invalid, BindBlock.EMPTY);
-      var asId = useAs.weakId().getText();
-      var asAssoc = useAs.assoc();
-      var asBind = useAs.bindBlock();
-      return new UseHide.Name(name, asId,
-        asAssoc != null ? visitAssoc(asAssoc) : Assoc.Invalid,
-        asBind != null ? visitBind(asBind) : BindBlock.EMPTY);
-    });
-  }
-
-  public @NotNull UseHide visitUseHide(@NotNull AyaParser.UseHideContext ctx) {
-    if (ctx.HIDING() != null) return hideList(ctx.hideList(), UseHide.Strategy.Hiding);
-    return useList(ctx.useList(), UseHide.Strategy.Using);
-  }
-
-  public @NotNull Command.Module visitModule(AyaParser.ModuleContext ctx) {
-    var id = ctx.weakId();
-    return new Command.Module(
-      sourcePosOf(id), sourcePosOf(ctx), id.getText(),
-      Seq.wrapJava(ctx.stmt()).flatMap(this::visitStmt)
-    );
-  }
-
-  public @NotNull Stream<WithPos<String>> visitIds(AyaParser.IdsContext ctx) {
-    return ctx.weakId().stream().map(id -> new WithPos<>(sourcePosOf(id), id.getText()));
-  }
-
-  public @NotNull Stream<String> visitIdsComma(AyaParser.IdsCommaContext ctx) {
-    return ctx.weakId().stream().map(ParseTree::getText);
-  }
-
-  public @NotNull Modifier visitFnModifiers(AyaParser.FnModifiersContext ctx) {
-    if (ctx.OPAQUE() != null) return Modifier.Opaque;
-    if (ctx.INLINE() != null) return Modifier.Inline;
-    if (ctx.OVERLAP() != null) return Modifier.Overlap;
-    return unreachable(ctx);
-  }
-
-  private @NotNull SourcePos sourcePosOf(ParserRuleContext ctx) {
-    return source.fold(sourceFile -> AntlrUtil.sourcePosOf(ctx, sourceFile), pos -> pos);
-  }
-
-  private @NotNull SourcePos sourcePosOf(TerminalNode node) {
-    return source.fold(sourceFile -> AntlrUtil.sourcePosOf(node, sourceFile), pos -> pos);
-  }
-}
