+++ cli/src/main/java/org/aya/cli/repl/jline/JlineRepl.java
@@ -13,9 +13,9 @@ import org.aya.pretty.doc.Doc;
 import org.aya.pretty.style.AyaStyleFamily;
 import org.aya.repl.CmdCompleter;
 import org.aya.repl.ReplUtil;
-import org.aya.repl.antlr.JFlexAdapter;
-import org.aya.repl.antlr.ReplHighlighter;
-import org.aya.repl.antlr.ReplParser;
+import org.aya.repl.gk.JFlexAdapter;
+import org.aya.repl.gk.ReplHighlighter;
+import org.aya.repl.gk.ReplParser;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.VisibleForTesting;
