+++ parser-migrate/build.gradle.kts
@@ -1,3 +1,5 @@
+// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
+// Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 val genDir = file("src/main/gen")
 idea.module.generatedSourceDirs.add(genDir)
 sourceSets.main {
