+++ tools-repl/src/main/java/module-info.java
@@ -5,10 +5,10 @@ module aya.repl {
   requires static org.jetbrains.annotations;
   requires transitive kala.base;
   requires transitive kala.collection;
-  requires org.antlr.antlr4.runtime;
+  requires aya.ij.parsing.core;
   requires org.jline.reader;
   requires org.jline.terminal;
 
-  exports org.aya.repl.antlr;
+  exports org.aya.repl.gk;
   exports org.aya.repl;
 }
