+++ tools-repl/src/main/java/org/aya/repl/gk/ReplParser.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.repl.antlr;
+package org.aya.repl.gk;
 
 import org.aya.repl.Command;
 import org.aya.repl.CommandArg;
